fixing frontend  saved chart reload
